+++ src/12/template/src/lib.rs
@@ -19,7 +19,7 @@ pub mod pallet {
 
 	#[pallet::storage]
 	/* 🚧 TODO 🚧: Update this storage to use a `QueryKind` of `ValueQuery`. */
-	pub(super) type CountForKitties<T: Config> = StorageValue<Value = u64>;
+	pub(super) type CountForKitties<T: Config> = StorageValue<Value = u32>;
 
 	#[pallet::event]
 	#[pallet::generate_deposit(pub(super) fn deposit_event)]
