+++ src/15/template/src/impls.rs
@@ -7,7 +7,7 @@ impl<T: Config> Pallet<T> {
 			- `ensure!` that `Kitties` map does not `contains_key` for `dna`.
 			- If it does, return `Error::<T>::DuplicateKitty`.
 		*/
-		let current_count: u64 = CountForKitties::<T>::get();
+		let current_count: u32 = CountForKitties::<T>::get();
 		let new_count = current_count.checked_add(1).ok_or(Error::<T>::TooManyKitties)?;
 		Kitties::<T>::insert(dna, ());
 		CountForKitties::<T>::set(new_count);
