+++ src/18/template/src/impls.rs
@@ -10,7 +10,7 @@ impl<T: Config> Pallet<T> {
 		// Check if the kitty does not already exist in our storage map
 		ensure!(!Kitties::<T>::contains_key(dna), Error::<T>::DuplicateKitty);
 
-		let current_count: u64 = CountForKitties::<T>::get();
+		let current_count: u32 = CountForKitties::<T>::get();
 		let new_count = current_count.checked_add(1).ok_or(Error::<T>::TooManyKitties)?;
 		/* 🚧 TODO 🚧: Insert `kitty`into the map instead of `()`. */
 		Kitties::<T>::insert(dna, ());
