+++ src/18/template/src/lib.rs
@@ -28,7 +28,7 @@ pub mod pallet {
 	}
 
 	#[pallet::storage]
-	pub(super) type CountForKitties<T: Config> = StorageValue<Value = u64, QueryKind = ValueQuery>;
+	pub(super) type CountForKitties<T: Config> = StorageValue<Value = u32, QueryKind = ValueQuery>;
 
 	#[pallet::storage]
 	/* 🚧 TODO 🚧: Update the `Value` to be type `Kitty<T>` instead of (). */
